(PC-29224)[PRO] refactor: move theme handling to App to prevent router to rerender
