+++ pro/src/app/AppRouter/AppRouter.tsx
@@ -1,11 +1,9 @@
 import * as Sentry from '@sentry/react'
-import { useEffect } from 'react'
 import { useSelector } from 'react-redux'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 
 import App from 'app/App/App'
 import routes from 'app/AppRouter/routesMap'
-import useIsNewInterfaceActive from 'hooks/useIsNewInterfaceActive'
 import { selectActiveFeatures } from 'store/features/selectors'
 
 import { ErrorBoundary } from './ErrorBoundary'
@@ -15,18 +13,10 @@ const sentryCreateBrowserRouter =
 
 const AppRouter = (): JSX.Element => {
   const activeFeatures = useSelector(selectActiveFeatures)
-  const isNewInterfaceActive = useIsNewInterfaceActive()
 
   const activeRoutes = routes.filter(
     (route) => !route.featureName || activeFeatures.includes(route.featureName)
   )
-
-  useEffect(() => {
-    document.documentElement.setAttribute(
-      'data-theme',
-      isNewInterfaceActive ? 'blue' : 'pink'
-    )
-  }, [isNewInterfaceActive])
 
   const router = sentryCreateBrowserRouter([
     {
